# Conflicts:
#	pom.xml
#	src/main/java/com/example/onlinebookstore/controller/BookController.java
#	src/main/java/com/example/onlinebookstore/service/impl/BookServiceImpl.java
#	src/main/resources/db/changelog/changes/01-create-book-table.yaml
#	src/main/resources/db/changelog/db.changelog-master.yaml

+++ src/main/java/com/example/onlinebookstore/service/impl/BookServiceImpl.java
@@ -33,17 +33,15 @@ public class BookServiceImpl implements BookService {
 
     @Override
     public BookDto findById(long id) {
-        Book book = bookRepository.findById(id).orElseThrow(
-                () -> new EntityNotFoundException("Book not found by id: " + id)
-        );
+        Book book = findBookByIdOrThrow(id);
         return bookMapper.toDto(book);
     }
 
     @Override
     public BookDto updateById(Long id, CreateBookRequestDto createBookRequestDto) {
-        bookRepository.findById(id).orElseThrow(
-                () -> new EntityNotFoundException("Book not found by id: " + id));
-
+        if (!validateExistence(id)) {
+            throw new EntityNotFoundException("Cant update book by id: " + id);
+        }
         Book book = bookMapper.toModel(createBookRequestDto);
         book.setId(id);
         Book savedBook = bookRepository.save(book);
@@ -52,6 +50,19 @@ public class BookServiceImpl implements BookService {
 
     @Override
     public void deleteById(Long id) {
+        if (!validateExistence(id)) {
+            throw new EntityNotFoundException("Cant delete book by id: " + id);
+        }
         bookRepository.deleteById(id);
+    }
+
+    private boolean validateExistence(Long id) {
+        return bookRepository.existsById(id);
+    }
+
+    private Book findBookByIdOrThrow(Long id) {
+        return bookRepository.findById(id).orElseThrow(
+                () -> new EntityNotFoundException("Book not found by id: " + id)
+        );
     }
 }
